Use ifuncs on Linux to avoid dispatch table overhead.
These tell the linker to generate GNU_IFUNC relocs that rewrite the
PLT entries in the user's address space to point to our resolved GL
function, so there's no extra function pointer.  It also, as a bonus,
cuts 400k out of the library.

This requires a toolchain from 2010 or so.  Unfortunately, it's going
to take a bit more investigation to find out what specific bits are
required.

Fixes #4