Add limited support for the KHRONOS_* defines from khrplatform.h.
The only way these get used is when you choose to include epoxy/gl.h
but EGL/egl.h.  I'm not sure that's something I want to actually
support, but piglit happened to do so, and this gets things working.

There's not as much support here as in real khrplatform.h, but then
there's no EGL for windows.  And seriously, compilers support 64-bit
and float types, why would you even build this complexity?