Add C99 Headers for use with MSVC (pre-MSVC 2013)
This adds C99 headers stdint.h, inttypes.h and stdbool.h that are not
available on Visual Studio 2012 and earlier, so that they can be referred
to during the build of libepoxy (and possibly the use of it).

The stdint.h and inttypes.h implementation are taken from the msinttypes
project[1], and the stdbool.h is adapted from the one from gnulib.

[1]: https://code.google.com/p/msinttypes/