Non-generated Sources: Allow Build on Visual Studio
As this code is obviously not only built on gcc, define a constructor that
is suitable for use on Visual Studio, in addition to the current gcc one.

Also, avoid defining an empty struct (which is not allowed on any MSVC),
and make sure that we have GLAPIENTRY (i.e. __stdcall) just before the
parameter part of the function declaration/prototype, as that is where
Visual Studio expects them to be, and is also accepted on gcc.

As epoxy_conservative_gl_version() is marked for export in
dispatch_common.c, do likewise for distpatch_common.h by using PUBLIC, as
Visual Studio does not like export annotations to be done on function
declarations but not the prototypes.