Makefile.am's: Move Source Listings into Makefile.sources's
This moves the source, headers and test program listings from the
respective Makefile.am's under src/ and test/ to their respective
Makefile.sources.  This is done so that we can have NMake Makefiles later
that can also refer to the same Makefile.sources to build the libepoxy DLL
and the relevant test programs.

src/Makefile.am has also been re-organized a bit so that we can ensure that
the non-generated sources and headers are distributed and built, but then
that the generated ones are not distributed.