MSVC Builds: Add NMake Makefiles
This adds NMake Makefiles that can be used to build the libepoxy DLL (with
WGL support), and the test programs that are relevant to the WGL "backend",
and adds the NMake Makefiles and the C99-compatibility headers needed for
pre-2013 MSVC versions to the distributions.  This NMake Makefile set
supports Visual Studio 2008 through 2013.

Like the earlier patch on the Makefile.am's, these NMake Makefiles also
consume the Makefile.sources to identify the sources needed for the build.