MSVC Builds: Don't Bundle C99 Compatilibilty Headers
Instead, update the NMake Makefiles and tell people in there that they will
need to go and acquire their own stdint.h, inttypes.h and stdbool.h to be
used for building and using libepoxy.  Also update the CFLAGS a bit, as
they can be grouped together, to make things neater.

Also add the copyright notice in msvc/detectenv-msvc.mak.