src/gen_dispatch.py: Make Generated Code Build Under C89
...so that the generated code are buildable by pre-2013 Visual Studio.  The
main thing that this does is that we avoid named initializers, but instead
initialize the structs in old-school C89 way.

The generated code may not look that robust, but since this is generated
code, I think this is not an issue; when the Khronos registry gets updated,
all that is needed is that the code gets re-generated, and we have the
items in the right order.

These are the only changes required, after getting the code to work with
Visual Studio 2013, for libepoxy to build with Visual Studio 2008-2012,
provided that the user finds a suitable implementation of stdint.h,
inttypes.h and stdbool.h.