nVidia glx.h uses __glx_h__; define that too.
The code used to check this symbol, but commit 689abf4 replaced it with
GLX_H, presumably to work better with either Mesa or Khronos headers.
But nVidia's header use the older include guard. Add it as another
option.

Should fix the headerguards.c compile test when the system glx.h is from
nVidia's binary drivers.

Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>