Add public API to query whether GLX support is available
Some environments may provide an X server, but no GLX extension;
usually, this is the case for heavily constrained systems, like an
installer.

Since Epoxy is in charge of getting through the various bits of GLX it
stands to reason that Epoxy should also provide a simple API for
checking whether or not GLX is available at run time; this allows
callers in projects depending on Epoxy, to bail out early and avoid
providing GL support altogether.