meson: generalise build_apple to has_dlvsym
build_apple was introduced in 756dca as a proxy for the fact that Apple's libc
doesn't have dlvsym(), which is glibc-specific so also isn't present in other
libc implementations such as musl.

Instead of detecting whether we are building for Apple or not, just probe the to
see if we have dlvsym.