build: Check for included system headers
The generated dispatch headers for EGL and GLX include system headers.
We do a pkg-config check for those, but the dependency is not required,
in order to support builds on systems that do not use pkg-config.

If the dependency is not found, we need to check for the headers we
include, otherwise building Epoxy will fail. If the support for EGL or
GLX was explicitly enabled, fail the configuration step; otherwise,
disable the EGL or GLX support.